Fix argparse pinning.
argparse was pinned to version 1.3 for python < 3.  argparse version
1.2.1 is built in to Python 2.7.  Version 1.3 adds alias support, which
is not used by gaia.  The pinned version doesn't install in a Python 2.7
virtualenv, resulting in the inability to import Gaia.

I've changed this so only Python 2.6 will import argparse.  Perhaps we
should remove support for Python 2.6, in which case we could remove
argparse as a requirement, since it is built in to Python 2.7 and
>=3.2.